docs(env): document config_root and update env._.path/source to use config_root\n\nfix(config): resolve env._.path relative to config_root; add process-local cache for config_root and reset it on Settings::reset()\n\nrefactor(config): move config_root logic into config_root.rs and update imports\n\n test(e2e): add test_env_path_resolution to verify path resolution against config_root
