+++ src/config/config_file/mod.rs
@@ -10,17 +10,15 @@ use std::{
 
 use eyre::{Result, eyre};
 use idiomatic_version::IdiomaticVersionFile;
-use path_absolutize::Absolutize;
 use serde_derive::Deserialize;
 use std::sync::LazyLock as Lazy;
 use tool_versions::ToolVersions;
 use versions::Versioning;
-use xx::regex;
 
 use crate::cli::args::{BackendArg, ToolArg};
 use crate::config::config_file::mise_toml::MiseToml;
 use crate::config::env_directive::EnvDirective;
-use crate::config::{AliasMap, Settings, is_global_config, settings};
+use crate::config::{AliasMap, Settings, settings};
 use crate::errors::Error::UntrustedConfig;
 use crate::file::display_path;
 use crate::hash::hash_to_str;
@@ -34,6 +32,7 @@ use crate::{backend, config, dirs, env, file, hash};
 
 use super::Config;
 
+pub mod config_root;
 pub mod idiomatic_version;
 pub mod mise_toml;
 pub mod toml;
@@ -72,7 +71,7 @@ pub trait ConfigFile: Debug + Send + Sync {
     }
     fn config_type(&self) -> ConfigFileType;
     fn config_root(&self) -> PathBuf {
-        config_root(self.get_path())
+        config_root::config_root(self.get_path())
     }
     fn plugins(&self) -> Result<HashMap<String, String>> {
         Ok(Default::default())
@@ -259,67 +258,11 @@ pub fn parse(path: &Path) -> Result<Arc<dyn ConfigFile>> {
     }
 }
 
-pub fn config_root(path: &Path) -> PathBuf {
-    if is_global_config(path) {
-        return env::MISE_GLOBAL_CONFIG_ROOT.to_path_buf();
-    }
-    let path = path
-        .absolutize()
-        .map(|p| p.to_path_buf())
-        .unwrap_or(path.to_path_buf());
-    let parts = path
-        .components()
-        .map(|c| c.as_os_str().to_string_lossy().to_string())
-        .collect::<Vec<_>>();
-    const EMPTY: &str = "";
-    let filename = parts.last().map(|p| p.as_str()).unwrap_or(EMPTY);
-    let parent = parts
-        .iter()
-        .nth_back(1)
-        .map(|p| p.as_str())
-        .unwrap_or(EMPTY);
-    let grandparent = parts
-        .iter()
-        .nth_back(2)
-        .map(|p| p.as_str())
-        .unwrap_or(EMPTY);
-    let great_grandparent = parts
-        .iter()
-        .nth_back(3)
-        .map(|p| p.as_str())
-        .unwrap_or(EMPTY);
-    let parent_path = || path.parent().unwrap().to_path_buf();
-    let grandparent_path = || parent_path().parent().unwrap().to_path_buf();
-    let great_grandparent_path = || grandparent_path().parent().unwrap().to_path_buf();
-    let great_great_grandparent_path = || great_grandparent_path().parent().unwrap().to_path_buf();
-    let is_mise_dir = |d: &str| d == "mise" || d == ".mise";
-    let is_config_filename = |f: &str| {
-        f == "config.toml" || f == "config.local.toml" || regex!(r"config\..+\.toml").is_match(f)
-    };
-    if parent == "conf.d" && is_mise_dir(grandparent) {
-        if great_grandparent == ".config" {
-            great_great_grandparent_path()
-        } else {
-            great_grandparent_path()
-        }
-    } else if is_mise_dir(parent) && is_config_filename(filename) {
-        if grandparent == ".config" {
-            great_grandparent_path()
-        } else {
-            grandparent_path()
-        }
-    } else if parent == ".config" {
-        grandparent_path()
-    } else {
-        parent_path()
-    }
-}
-
 pub fn config_trust_root(path: &Path) -> PathBuf {
     if settings::is_loaded() && Settings::get().paranoid {
         path.to_path_buf()
     } else {
-        config_root(path)
+        config_root::config_root(path)
     }
 }
 
@@ -610,36 +553,5 @@ mod tests {
             detect_config_file_type(Path::new("/foo/bar/rust-toolchain.toml")),
             Some(ConfigFileType::IdiomaticVersion)
         );
-    }
-
-    #[test]
-    fn test_config_root() {
-        for p in &[
-            "/foo/bar/.config/mise/conf.d/config.toml",
-            "/foo/bar/.config/mise/conf.d/foo.toml",
-            "/foo/bar/.config/mise/config.local.toml",
-            "/foo/bar/.config/mise/config.toml",
-            "/foo/bar/.config/mise.local.toml",
-            "/foo/bar/.config/mise.toml",
-            "/foo/bar/.mise.env.toml",
-            "/foo/bar/.mise.local.toml",
-            "/foo/bar/.mise.toml",
-            "/foo/bar/.mise/conf.d/config.toml",
-            "/foo/bar/.mise/config.local.toml",
-            "/foo/bar/.mise/config.toml",
-            "/foo/bar/.tool-versions",
-            "/foo/bar/mise.env.toml",
-            "/foo/bar/mise.local.toml",
-            "/foo/bar/mise.toml",
-            "/foo/bar/mise/config.local.toml",
-            "/foo/bar/mise/config.toml",
-            "/foo/bar/.config/mise/config.env.toml",
-            "/foo/bar/.config/mise.env.toml",
-            "/foo/bar/.mise/config.env.toml",
-            "/foo/bar/.mise.env.toml",
-        ] {
-            println!("{p}");
-            assert_eq!(config_root(Path::new(p)), PathBuf::from("/foo/bar"));
-        }
     }
 }
