+++ src/config/env_directive/mod.rs
@@ -1,4 +1,4 @@
-use crate::config::config_file::{config_root, trust_check};
+use crate::config::config_file::trust_check;
 use crate::dirs;
 use crate::env;
 use crate::env_diff::EnvMap;
@@ -217,7 +217,7 @@ impl EnvResults {
             //     &directive,
             //     &source
             // );
-            let config_root = config_root(&source);
+            let config_root = crate::config::config_file::config_root::config_root(&source);
             ctx.insert("cwd", &*dirs::CWD);
             ctx.insert("config_root", &config_root);
             let env_vars = env
@@ -360,11 +360,9 @@ impl EnvResults {
         // trace!("resolve: paths: {:#?}", &paths);
         // trace!("resolve: ctx.env: {:#?}", &ctx.get("env"));
         for (source, paths) in &paths.iter().chunk_by(|(_, source)| source) {
-            let config_root = source
-                .parent()
-                .map(Path::to_path_buf)
-                .or_else(|| dirs::CWD.clone())
-                .unwrap_or_default();
+            // Use the computed config_root (project root for nested configs) for path resolution
+            // to be consistent with other env directives like _.source and _.file
+            let config_root = crate::config::config_file::config_root::config_root(source);
             let paths = paths.map(|(p, _)| p).collect_vec();
             let mut paths = paths
                 .iter()
